refactor(settings): route all access through @Dependency(\.settings)
Hide `Settings.shared` behind `fileprivate` and inline the helper that
used to live in `Settings+DependencyValues.swift` into `Settings.swift`,
so every caller must resolve the singleton via `@Dependency(\.settings)`
(or an override) instead of touching the static accessor directly.
Update the two remaining call sites in
`RuntimeConnectionNotificationService` and `MainViewModel` accordingly.
